+++ bin/domain/imag-calendar/src/collection.rs
@@ -17,7 +17,6 @@
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 //
 
-use std::ops::Deref;
 use std::path::PathBuf;
 use std::process::exit;
 use std::io::Write;
@@ -27,7 +26,6 @@ use filters::filter::Filter;
 use walkdir::WalkDir;
 use walkdir::DirEntry;
 use clap::ArgMatches;
-use prettytable::Table;
 use itertools::Itertools;
 
 use libimagerror::exit::ExitUnwrap;
@@ -40,10 +38,8 @@ use libimagcalendar::store::collections::CalendarCollectionStore;
 use libimagentryref::reference::Ref;
 use libimagcalendar::collection::Collection;
 use libimagstore::iter::get::StoreIdGetIteratorExtension;
-use libimagstore::store::FileLockEntry;
 use libimagutil::warn_result::*;
 use libimagcalendar::calendar::Calendar;
-use libimagcalendar::event::Event;
 
 use util::{GrepFilter, PastFilter};
 
@@ -233,7 +229,7 @@ fn show<'a>(rt: &Runtime, scmd: &ArgMatches<'a>) {
         .flatten()
         .filter(|e| past_filter.filter(e));
 
-    show_events(rt, iterator);
+    ::util::show_events(rt, iterator);
 }
 
 fn list<'a>(rt: &Runtime, scmd: &ArgMatches<'a>) {
@@ -267,7 +263,7 @@ fn list<'a>(rt: &Runtime, scmd: &ArgMatches<'a>) {
         .flatten()
         .filter(|f| past_filter.filter(f));
 
-    list_events(rt, scmd.is_present("collection-list-table"), iterator);
+    ::util::list_events(rt, scmd.is_present("collection-list-table"), iterator);
 }
 
 fn find<'a>(rt: &Runtime, scmd: &ArgMatches<'a>) {
@@ -310,117 +306,9 @@ fn find<'a>(rt: &Runtime, scmd: &ArgMatches<'a>) {
         .filter(|e| filter.filter(e));
 
     if do_show {
-        show_events(rt, iterator);
+        ::util::show_events(rt, iterator);
     } else {
-        list_events(rt, scmd.is_present("collection-find-table"), iterator);
-    }
-}
-
-
-//
-// Helpers
-//
-
-fn show_events<'a, I>(rt: &Runtime, iter: I)
-    where I: Iterator<Item = FileLockEntry<'a>>
-{
-    let out           = rt.stdout();
-    let mut outlock   = out.lock();
-    let get_show_data = |event: &FileLockEntry| {
-            let start = event
-                .get_start()
-                .map_err_trace_exit_unwrap(1)
-                .format(::libimagtimeui::ui::time_ui_fmtstr());
-
-            let end = event
-                .get_end()
-                .map_err_trace_exit_unwrap(1)
-                .format(::libimagtimeui::ui::time_ui_fmtstr());
-
-            let desc = event
-                .get_description()
-                .map_err_trace_exit_unwrap(1);
-
-            let cats = event
-                .get_categories()
-                .map_err_trace_exit_unwrap(1);
-
-            let loca = Event::get_location(event.deref())
-                .map_err_trace_exit_unwrap(1);
-
-            (start, end, desc, cats, loca)
-    };
-
-    iter.for_each(|event| {
-        let (s, e, d, c, l) = get_show_data(&event);
-        let c               = c.join(", "); // join categories by ", "
-        let hash            = event.get_hash().map_err_trace_exit_unwrap(1);
-
-        let _ = writeln!(outlock,
-                         r#"Event Id   : {hash}
-                            Start      : {start}
-                            End        : {end}
-                            Description: {description}
-                            Categories : {categories}
-                            Location   : {location}
-                            "#,
-                            hash        = hash,
-                            start       = s,
-                            end         = e,
-                            description = d,
-                            categories  = c,
-                            location    = l)
-            .to_exit_code()
-            .unwrap_or_exit();
-    });
-}
-
-fn list_events<'a, I>(rt: &Runtime, table: bool, iter: I)
-    where I: Iterator<Item = FileLockEntry<'a>>
-{
-    let out           = rt.stdout();
-    let mut outlock   = out.lock();
-    let get_list_data = |event: &FileLockEntry| {
-            let start = event
-                .get_start()
-                .map_err_trace_exit_unwrap(1)
-                .format(::libimagtimeui::ui::time_ui_fmtstr());
-
-            let end = event
-                .get_end()
-                .map_err_trace_exit_unwrap(1)
-                .format(::libimagtimeui::ui::time_ui_fmtstr());
-
-            let desc = event
-                .get_description()
-                .map_err_trace_exit_unwrap(1);
-
-            (start, end, desc)
-    };
-
-    if table {
-        let mut tab = Table::new();
-        tab.add_row(row!["Start", "End", "Description"]);
-
-        iter.for_each(|event| {
-            let (start, end, desc) = get_list_data(&event);
-            tab.add_row(row![start, end, desc]);
-        });
-
-        let _ = tab.print(&mut out.lock())
-            .unwrap_or_else(|e| {
-                error!("IO error: {:?}", e);
-                exit(1)
-            });
-    } else {
-        iter.for_each(|event| {
-            let (start, end, desc) = get_list_data(&event);
-            let hash               = event.get_hash().map_err_trace_exit_unwrap(1);
-
-            let _ = writeln!(outlock, "{}: {} - {} - {}", hash, start, end, desc)
-                .to_exit_code()
-                .unwrap_or_exit();
-        });
+        ::util::list_events(rt, scmd.is_present("collection-find-table"), iterator);
     }
 }
